[Practica 01]: Todo listo para evaluación, sistema de logs implementado y ejecución mejorada para visualización del usuario

+++ Practica01/include/tools.h
@@ -17,7 +17,7 @@
 #include <fstream>
 #include "types.h"
 
-int CheckParameters(const int argc, char *argv[], int& size, Frontier& frontier_type, std::string& input_file);
+int CheckParameters(const int argc, char *argv[], long unsigned int& size, Frontier& frontier_type, std::string& input_file);
 void Usage(const bool mode);
 bool IsDigit(const std::string& str);
 std::string ExtractInitialConfiguration(const std::string& input_file);
